% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_calc.R
\name{age_calc}
\alias{age_calc}
\title{Age calucaltion function}
\usage{
age_calc(dob, enddate = Sys.Date(), units = "months")
}
\description{
param dob date of birth
param enddate date to calculate difference from
param units "days", "months", "years"
}
