% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbetsfri_dag.R
\name{arbetsfri_dag}
\alias{arbetsfri_dag}
\title{Get arbetsfria dagar}
\usage{
arbetsfri_dag(year)
}
\arguments{
\item{year}{year that you want to know arbetsfria dagar for}
}
\description{
Get arbetsfria dagar
}
\examples{
\dontrun{
arbetsfri_dag(2019)
}
}
